Energy Refill Update
It was observed that if we keep wait time to refill as 1 min and we wait 30 sec in game and then close the app. When we reopen after some time, it again waits for remaining time(30 sec) and then play button will get enables. This issue has been overcome by use of OnApplicationFocus() method.